requirements.txt
Django==3.1.3
tensorflow==2.1.0
Keras==2.3.1
Pillow==6.2.0
numpy==1.16.5
h5py==2.9.0